Now can be return items by query

+++ backend/MercadoLibre.Backend/MercadoLibre.Backend.Domain/Responses/PriceResponse.cs
@@ -11,6 +11,6 @@ namespace MercadoLibre.Backend.Domain.Responses
         public decimal Amount { get; set; }
 
         [JsonProperty(PropertyName = "decimals")]
-        public int Decimals { get; set; }
+        public decimal Decimals { get; set; }
     }
 }
